Large refactor

+++ BugTracker/Controllers/ProjectsController.cs
@@ -19,7 +19,7 @@ namespace BugTracker.Controllers
 {
 	public class ProjectsController : Controller
 	{
-		private readonly ILogger<HomeController> _logger;
+		private readonly ILogger<HomeController> logger;
 		private readonly IProjectRepository projectRepository;
 		private readonly IBugReportRepository bugReportRepository;
 		private readonly IActivityRepository activityRepository;
@@ -37,7 +37,7 @@ namespace BugTracker.Controllers
 									IProjectInviter projectInvitation,
 									ApplicationUserManager userManager)
 		{
-			this._logger = logger;
+			this.logger = logger;
 			this.projectRepository = projectRepository;
 			this.bugReportRepository = bugReportRepository;
 			this.activityRepository = activityRepository;
@@ -53,12 +53,8 @@ namespace BugTracker.Controllers
 			var model = await projectRepository.GetAll();
 				model = model.Where(project =>
 					{
-						Task<AuthorizationResult> authorizationResult = authorizationService.AuthorizeAsync(HttpContext.User, project.ProjectId, "CanAccessProjectPolicy");
-						if(authorizationResult.IsCompletedSuccessfully && authorizationResult.Result.Succeeded)
-						{
-							return true;
-						}
-						else { return false; }
+						var authorizationResult = authorizationService.AuthorizeAsync(HttpContext.User, project.ProjectId, "CanAccessProjectPolicy");
+						return authorizationResult.IsCompletedSuccessfully && authorizationResult.Result.Succeeded;
 					}
 				);
 
@@ -68,28 +64,28 @@ namespace BugTracker.Controllers
 		public async Task<IActionResult> Overview(int id)
 		{
 			var authorizationResult = authorizationService.AuthorizeAsync(HttpContext.User, id, "CanAccessProjectPolicy");
-			if(authorizationResult.IsCompletedSuccessfully && authorizationResult.Result.Succeeded)
+			if (!authorizationResult.IsCompletedSuccessfully || authorizationResult.Result.Succeeded)
 			{
-				Project project = await projectRepository.GetById(id);
-				HttpContext.Session.SetInt32("currentProject", id); // save project id to session
 
-				IEnumerable<BugReport> bugReports = await bugReportRepository.GetAllById(id);
-
-				OverviewProjectViewModel overviewProjectViewModel = new OverviewProjectViewModel()
-				{
-					Project = project,
-					BugReports = bugReports.ToList(),
-					CommentCountHandler = bugReportRepository.GetCommentCountById
-				};
-
-				ViewData["BreadcrumbNode"] = BreadcrumbNodeHelper.ProjectOverview(project);
-				
-				// if project NULL -- redirect to error page !!
-
-				return View(overviewProjectViewModel);
+				return RedirectToAction("Index", "Home");
 			}
+			
+			var project = await projectRepository.GetById(id);
+			if (project == null) return NotFound();
+			HttpContext.Session.SetInt32("currentProject", id); // save project id to session
 
-			return RedirectToAction("Index", "Home");
+			var bugReports = await bugReportRepository.GetAllById(id);
+
+			var viewModel = new OverviewProjectViewModel()
+			{
+				Project = project,
+				BugReports = bugReports.ToList(),
+				CommentCountHandler = bugReportRepository.GetCommentCountById
+			};
+
+			ViewData["BreadcrumbNode"] = BreadcrumbNodeHelper.ProjectOverview(project);
+
+			return View(viewModel);
 		}
 
 		[HttpGet]
@@ -104,45 +100,49 @@ namespace BugTracker.Controllers
 		[Authorize]
 		public async Task<IActionResult> CreateProject(Project model)
 		{
-			if (ModelState.IsValid)
+			if (!ModelState.IsValid)
 			{
-				Project newProject = new Project
-				{
-					Name = model.Name,
-					Description = model.Description,
-					CreationTime = DateTime.Now,
-					LastUpdateTime = DateTime.Now,
-					Hidden = model.Hidden,
-					BugReports = new List<BugReport>()
-				};
-
-				Project addedProject = await projectRepository .Add(newProject);
-				await bugReportRepository.AddLocalBugReportId(addedProject.ProjectId);
-
-				// Create activity event
-				int userId = Int32.Parse(httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value);
-				var activityEvent = new ActivityProject(DateTime.Now, addedProject.ProjectId, ActivityMessage.ProjectCreated, userId);
-				await activityRepository.Add(activityEvent);
-
-				// Add the user who created the project to its administrator role
-				var user = await userManager.FindByIdAsync(userId.ToString());
-				await userManager.AddToRoleAsync(user, "Administrator", addedProject.ProjectId);
-
-				_logger.LogInformation($"New project created. ID: {addedProject.ProjectId}, Name: {addedProject.Name}");
-
-				return RedirectToAction("Overview", new { id = addedProject.ProjectId });
+				return View();
 			}
+			
+			var project = new Project
+			{
+				Name = model.Name,
+				Description = model.Description,
+				CreationTime = DateTime.Now,
+				LastUpdateTime = DateTime.Now,
+				Hidden = model.Hidden,
+				BugReports = new List<BugReport>()
+			};
 
-			return View();
+			_ = await projectRepository.Add(project);
+			await bugReportRepository.AddLocalBugReportId(project.ProjectId);
+
+			// Create activity event
+			var userId = httpContextAccessor.HttpContext?.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+			
+			var activityEvent = new ActivityProject(DateTime.Now, project.ProjectId, ActivityMessage.ProjectCreated, userId);
+			await activityRepository.Add(activityEvent);
+
+			// Add the user who created the project to its administrator role
+			var user = await userManager.FindByIdAsync(userId);
+			await userManager.AddToRoleAsync(user, "Administrator", project.ProjectId);
+
+			logger.LogInformation($"New project created. ID: {project.ProjectId}, Name: {project.Name}");
+
+			return RedirectToAction("Overview", new { id = project.ProjectId });
 		}
 
-		public IActionResult DeleteProject(int id)
+		public async Task<IActionResult> DeleteProject(int id)
 		{
-			var authorizationResult = authorizationService.AuthorizeAsync(HttpContext.User, id, "ProjectAdministratorPolicy");
+			var user = httpContextAccessor.HttpContext?.User;
+			if (user is null) return BadRequest();
+			
+			var authorizationResult = authorizationService.AuthorizeAsync(user, id, "ProjectAdministratorPolicy");
 			if (authorizationResult.IsCompletedSuccessfully && authorizationResult.Result.Succeeded)
 			{
-				projectRepository.Delete(id);
-				_logger.LogInformation($"Project deleted. ID: {id}");
+				await projectRepository.Delete(id);
+				logger.LogInformation($"Project deleted. ID: {id}");
 			}
 
 			return RedirectToAction("Projects");
@@ -151,92 +151,113 @@ namespace BugTracker.Controllers
 		[HttpGet]
 		public async Task<IActionResult> Edit(int id)
 		{
-			var authorizationResult = authorizationService.AuthorizeAsync(HttpContext.User, id, "ProjectAdministratorPolicy");
-			if (authorizationResult.IsCompletedSuccessfully && authorizationResult.Result.Succeeded)
+			var user = httpContextAccessor.HttpContext?.User;
+			if (user is null) return BadRequest();
+			
+			var authorizationResult = authorizationService.AuthorizeAsync(user, id, "ProjectAdministratorPolicy");
+			if (!authorizationResult.IsCompletedSuccessfully || !authorizationResult.Result.Succeeded)
 			{
-				var project = await projectRepository.GetById(id);
-				EditProjectViewModel projectViewModel = new EditProjectViewModel
-				{
-					Project = project
-				};
-
-				ViewData["BreadcrumbNode"] = BreadcrumbNodeHelper.ProjectEdit(project);
-				
-				return View(projectViewModel);
+				return RedirectToAction("Overview", new { id });
 			}
-			return RedirectToAction("Overview", new { id });
+			
+			var project = await projectRepository.GetById(id);
+			var viewModel = new EditProjectViewModel
+			{
+				Project = project
+			};
+
+			ViewData["BreadcrumbNode"] = BreadcrumbNodeHelper.ProjectEdit(project);
+			
+			return View(viewModel);
 		}
 
 		[HttpPost]
-		public async Task<IActionResult> Edit(EditProjectViewModel model)
+		public async Task<IActionResult> Edit(EditProjectViewModel viewModel)
 		{
-			var authorizationResult = authorizationService.AuthorizeAsync(HttpContext.User, model.Project.ProjectId, "ProjectAdministratorPolicy");
-			if (authorizationResult.IsCompletedSuccessfully && authorizationResult.Result.Succeeded)
+			var user = httpContextAccessor.HttpContext?.User;
+			if (user is null) return BadRequest();
+			
+			var authorizationResult = authorizationService.AuthorizeAsync(user, viewModel.Project.ProjectId, "ProjectAdministratorPolicy");
+			if (!authorizationResult.IsCompletedSuccessfully || !authorizationResult.Result.Succeeded)
 			{
-				if (ModelState.IsValid)
-				{
-					Project project = await projectRepository .GetById(model.Project.ProjectId);
-					project.Name = model.Project.Name;
-					project.Description = model.Project.Description;
-					project.Hidden = model.Project.Hidden;
-					project.LastUpdateTime = DateTime.Now;
-
-					_ = projectRepository.Update(project);
-
-					// Create activity event
-					int userId = Int32.Parse(httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value);
-					var activityEvent = new ActivityProject(DateTime.Now, model.Project.ProjectId, ActivityMessage.ProjectEdited, userId);
-					await activityRepository .Add(activityEvent);
-
-					return RedirectToAction("Overview", new { id = project.ProjectId });
-				}
+				return Forbid();
 			}
 
-			return View();
+			if (!ModelState.IsValid)
+			{
+				return View(viewModel);
+			}
+			
+			var project = await projectRepository.GetById(viewModel.Project.ProjectId);
+			if (project == null) return BadRequest();
+			
+			project.Name = viewModel.Project.Name;
+			project.Description = viewModel.Project.Description;
+			project.Hidden = viewModel.Project.Hidden;
+			project.LastUpdateTime = DateTime.Now;
+
+			_ = projectRepository.Update(project);
+
+			// Create activity event
+			var userId = httpContextAccessor.HttpContext?.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+			if (userId is null) return BadRequest();
+			
+			var activityEvent = new ActivityProject(DateTime.Now, viewModel.Project.ProjectId, ActivityMessage.ProjectEdited, userId);
+			await activityRepository .Add(activityEvent);
+
+			return RedirectToAction("Overview", new { id = project.ProjectId });
 		}
 
 		[HttpGet]
 		public async Task<IActionResult> Invites(int id)
 		{
-			var authorizationResult = authorizationService.AuthorizeAsync(HttpContext.User, id, "ProjectAdministratorPolicy");
-			if (authorizationResult.IsCompletedSuccessfully && authorizationResult.Result.Succeeded)
+			var user = httpContextAccessor.HttpContext?.User;
+			if (user is null) return BadRequest();
+			
+			var authorizationResult = authorizationService.AuthorizeAsync(user, id, "ProjectAdministratorPolicy");
+			if (!authorizationResult.IsCompletedSuccessfully || !authorizationResult.Result.Succeeded)
 			{
-				var project = await projectRepository.GetById(id);
-				InvitesViewModel invitesViewModel = new InvitesViewModel
-				{
-					ProjectId = id
-				};
-
-				ViewData["BreadcrumbNode"] = BreadcrumbNodeHelper.ProjectInvites(project);
-				
-				return View(invitesViewModel);
+				return RedirectToAction("Overview", id);
 			}
+			
+			var project = await projectRepository.GetById(id);
+			var viewModel = new InvitesViewModel
+			{
+				ProjectId = id
+			};
 
-			return RedirectToAction("Overview", id);
+			ViewData["BreadcrumbNode"] = BreadcrumbNodeHelper.ProjectInvites(project);
+			
+			return View(viewModel);
 		}
 
 		[HttpPost]
 		public async Task<IActionResult> Invites(InvitesViewModel model)
 		{
-			var authorizationResult = authorizationService.AuthorizeAsync(HttpContext.User, model.ProjectId, "ProjectAdministratorPolicy");
-			if (authorizationResult.IsCompletedSuccessfully && authorizationResult.Result.Succeeded)
+			var user = httpContextAccessor.HttpContext?.User;
+			if (user is null) return BadRequest();
+			
+			var authorizationResult = authorizationService.AuthorizeAsync(user, model.ProjectId, "ProjectAdministratorPolicy");
+			if (!authorizationResult.IsCompletedSuccessfully || !authorizationResult.Result.Succeeded)
 			{
-				if (ModelState.IsValid)
-				{
-					ProjectInvitation invitation = new ProjectInvitation
-					{
-						EmailAddress = model.EmailAddress,
-						Project = await projectRepository .GetById(model.ProjectId),
-						ToUser = null,
-						FromUser = await userManager.GetUserAsync(HttpContext.User)
-					};
-
-					await projectInviter.AddProjectInvitation(invitation);
-					return RedirectToAction("Overview", "Projects", new { id = model.ProjectId });
-				}
+				return Forbid();
 			}
 
-			return View(model);
+			if (!ModelState.IsValid)
+			{
+				return View(model);
+			}
+			
+			var invitation = new ProjectInvitation
+			{
+				EmailAddress = model.EmailAddress,
+				Project = await projectRepository .GetById(model.ProjectId),
+				ToUser = null,
+				FromUser = await userManager.GetUserAsync(HttpContext.User)
+			};
+
+			await projectInviter.AddProjectInvitation(invitation);
+			return RedirectToAction("Overview", "Projects", new { id = model.ProjectId });
 		}
 	}
 }
